feat: 공통으로 사용되는 FrequencyInfo 컴포넌트 생성 후 적용
