--- src/app/(team)/team/[teamid]/task/[taskid]/_components/TaskDetailSection/FrequencyInfo.tsx
@@ -1,3 +1,0 @@
-export default function FrequencyInfo() {
-  return <div>주기</div>;
-}
