+++ src/app/(team)/team/[teamid]/task/[taskid]/_components/TaskDetailSection/index.tsx
@@ -2,7 +2,7 @@ import { TaskResponse } from '@/lib/apis/task/type';
 import TaskMenuButton from '@/components/common/task/TaskMenuButton';
 import WriterInfo from '@/components/common/user/WriterInfo';
 import DateInfo from '@/app/(team)/team/[teamid]/task/[taskid]/_components/TaskDetailSection/DateInfo';
-import FrequencyInfo from '@/app/(team)/team/[teamid]/task/[taskid]/_components/TaskDetailSection/FrequencyInfo';
+import FrequencyInfo from '@/components/common/task/FrequencyInfo';
 
 export default function TaskDetailSection({
   id,
@@ -25,7 +25,7 @@ export default function TaskDetailSection({
             {writer && <WriterInfo {...writer} />}
             <div className="flex gap-2.5">
               <DateInfo />
-              <FrequencyInfo />
+              <FrequencyInfo frequency={frequency} />
             </div>
           </div>
           <div className="text-md-regular">{description}</div>
