+++ src/app/(team)/team/[teamid]/tasklist/_components/TaskListSection/TaskCard/index.tsx
@@ -3,6 +3,7 @@ import { formatDate } from '@/utils/formatDate';
 import TitleButton from '@/app/(team)/team/[teamid]/tasklist/_components/TaskListSection/TaskCard/TitleButton';
 import IconRenderer from '@/components/common/Icons/IconRenderer';
 import TaskMenuButton from '@/components/common/task/TaskMenuButton';
+import FrequencyInfo from '@/components/common/task/FrequencyInfo';
 
 export default function TaskCard({
   id,
@@ -13,11 +14,6 @@ export default function TaskCard({
   commentCount,
 }: TaskResponse) {
   const formattedDate = formatDate(date);
-  const frequencyText: Record<string, string> = {
-    DAILY: '매일 반복',
-    WEEKLY: '매주 반복',
-    MONTHLY: '매월 반복',
-  };
 
   return (
     <div className="flex flex-col gap-2.5 rounded-lg bg-slate-800 px-[14px] py-3">
@@ -43,17 +39,7 @@ export default function TaskCard({
           <IconRenderer name="CalendarIcon" size={16} />
           <div className="text-xs-regular text-slate-500">{formattedDate}</div>
         </div>
-        {frequency !== 'ONCE' && (
-          <>
-            <div className="h-2 w-px bg-slate-700" />
-            <div className="flex items-center gap-1.5">
-              <IconRenderer name="RepeatIcon" size={16} />
-              <div className="text-xs-regular text-slate-500">
-                {frequencyText[frequency]}
-              </div>
-            </div>
-          </>
-        )}
+        <FrequencyInfo frequency={frequency} />
       </div>
     </div>
   );
